Changed DataType from Int to Long

+++ src/Adder.java
@@ -1,5 +1,5 @@
 public class Adder { 
-  public int add(int num1,int num2) { 
+  public long add(long num1,long num2) { 
      return num1+num2;
